test(state_store): extract shared LMDB test-store helper
Move make_test_store out of app_store's test module into a crate-visible test_support module so upcoming inspect tests can construct in-process stores without duplicating the env setup.
